requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asttokens 
async-timeout==4.0.3
attrs==23.2.0
backcall
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
cchardet==2.1.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama 
comm 
courlan==0.9.5
cssselect==1.2.0
datasets==2.16.1
dateparser==1.2.0
debugpy 
decorator 
dill==0.3.7
distro==1.9.0
exceptiongroup==1.2.0
executing 
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
htmldate==1.7.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
importlib-metadata 
ipykernel 
ipython 
jedi 
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client 
jupyter_core 
jusText==3.0.0
langcodes==3.3.0
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib-inline 
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest_asyncio 
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.3
openai==1.8.0
packaging 
pandas==2.1.4
parso 
pickleshare 
pillow==10.2.0
prompt-toolkit 
proto-plus==1.23.0
protobuf==4.25.2
psutil 
pure-eval
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments 
python-dateutil 
python-dotenv==1.0.1
pytz==2023.3.post1
pywin32==227
PyYAML==6.0.1
pyzmq 
rank-bm25==0.2.2
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.1
sgmllib3k==1.0.0
six 
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data 
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
tinysegmenter==0.3
tld==0.13
tldextract==5.1.1
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
tornado
tqdm==4.66.1
trafilatura==1.6.4
traitlets 
transformers==4.36.2
typing_extensions 
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth 
xxhash==3.4.1
yarl==1.9.4
zipp
